# swagger/ft.yaml
openapi: 3.0.0
info:
  title: ftx
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
...
  /api/business-idea/:
    put:
      tags: [ business-idea ]
      summary: Create a business-idea
      operationId: put-api-business-idea
      description: Create a business-idea
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/BusinessIdeaCreation'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
  /api/business-idea/{businessIdeaId}:
    post:
      tags: [ business-idea ]
      summary: Update a business-idea
      operationId: post-api-business-idea-{businessIdeaId}
      description: Update a business-idea by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: businessIdeaId
          in: 'path'
          description: "businessIdeaId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/BusinessIdeaUpdation'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
    get:
      tags: [ business-idea ]
      summary: Fetch the business-idea
      operationId: get-api-business-idea-{businessIdeaId}
      description: Fetch the business-idea by Id
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: businessIdeaId
          in: 'path'
          description: "businessIdeaId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/BusinessIdea'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
...
  /api/business-idea/{businessIdeaId}/pitches:
    get:
      tags: [ business-idea ]
      summary: Retrieves existing business idea
      operationId: get-api-business-idea-{businessIdeaId}-pitches
      description: Retrieves existing business idea
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: businessIdeaId
          in: 'path'
          description: "businessIdeaId"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/BusinessIdea'
                  - $ref: '#/components/schemas/GenericErrors'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - API Key: [ ]
  '/api/business-idea/logo/{businessIdeaId}':
    post:
      summary: Upload logo image by businessIdeaId.
      tags: [ business-idea ]
      operationId: post-business-logo
      description: Upload logo image by businessIdeaId.
      parameters:
        - name: "AuthToken"
          in: 'header'
          description: "Authorization token"
          required: true
          schema:
            type: "string"
        - name: "businessIdeaId"
          in: 'path'
          description: "The business idea ID to upload image to"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BusinessIdeaUploadLogoReq'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/BusinessIdea'
                  - $ref: '#/components/schemas/GenericErrors'
        '422':
          description: Invalid Argument
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                    example:
                      - title: "Invalid value"
        '500':
          description: Internal Server Error
...
components:
...
  schemas:
...
    BusinessIdea:
      properties:
        userId:
          type: object
          $ref: '#/components/schemas/UserProfileUserEditable'
        latestAvgPitchRating:
          type: number
        title:
          type: string
        pitchCount:
          type: number
          default: 0
        description:
          type: string
        language:
          type: string
        logo:
          type: string
        websiteUrl:
          type: string
        industries:
          type: array
          items:
            type: string
        location:
          type: array
          items:
            type: string
        stage:
          type: string
...
    BusinessIdeaCreation:
      description: ''
      type: object
      x-examples:
        example-1:
          title: string
          description: string
          language: string
          logo: string
          websiteUrl: string
          industries: array
          location: array
          stage: string
      title: BusinessIdeaCreation
      properties:
        title:
          type: string
          description: business-idea's title
          example: Best business idea
        description:
          type: string
          description: business-idea's description
          example: Greatest business idea ever
        language:
          type: string
          description: business-idea's language in ISO 639-1 format
          example: en
        logo:
          type: string
          description: business-idea's logo link
          example: https://logoforexample.com
        websiteUrl:
          type: string
          description: business-idea's website url
          example: https://websiteforexample.com
        industries:
          type: array
          description: bysiness-idea's industries
          example: ['industries 1','industries 2']
        location:
          type: array
          description: bysiness-idea's locations
          example: ['location 1','location 2']
        stage:
          type: string
          description: bysiness-idea's stage
          example: stage 1
    BusinessIdeaUpdation:
      description: ''
      type: object
      x-examples:
        example-1:
          title: string
          description: string
          language: string
          logo: string
          websiteUrl: string
          industries: array
          location: array
          stage: string
      title: BusinessIdeaUpdation
      properties:
        title:
          type: string
          description: new business-idea's title
          example: Updated best business idea
        description:
          type: string
          description: new business-idea's description
          example: Updated greatest business idea ever
        language:
          type: string
          description: new business-idea's language in ISO 639-1 format
          example: fr
        logo:
          type: string
          description: new business-idea's logo link
          example: https://updatedlogoforexample.com
        websiteUrl:
          type: string
          description: new business-idea's website url
          example: https://updatedwebsiteforexample.com
        industries:
          type: array
          description: new bysiness-idea's industries
          example: ['updated industries']
        location:
          type: array
          description: new bysiness-idea's locations
          example: ['updated location']
        stage:
          type: string
          description: new bysiness-idea's stage
          example: updated stage
...
    UserProfileUserEditable:
      description: ''
      type: object
      x-examples:
        example-1:
          email: string
          name: string
          surname: string
          avatarUrl: string
          bio: string
          theme: string
          watchedVideos: 0
      title: UserProfileUserEditable
      properties:
        email:
          type: string
          minLength: 1
          description: User email. Should be unique across the users
          example: test@domain.com
        name:
          type: string
          minLength: 1
          description: User first name
          example: BoJack
        surname:
          type: string
          minLength: 1
          description: User last name
          example: Horseman
        avatarUrl:
          type: string
          minLength: 1
          description: User's avatar URL
          example: 'https://i.pravatar.cc/300'
        bio:
          type: string
          minLength: 1
          description: User's bio
          example: User's bio goes here
        theme:
          type: string
          minLength: 1
          description: User's client application theme
          enum:
            - light
            - dark
...
    BusinessIdeaUploadLogoReq:
      description: 'Object sent from Front-end for Update business idea logo'
      type: object
      title: BusinessIdeaUploadLogoReq
      properties:
        image:
          type: object
          description: The logo file.
          properties:
            file:
              type: string
              format: binary
    GenericErrors:
      title: GenericErrors
      type: object
      x-examples:
        example-1:
          - string
      properties:
        errors:
          type: array
          description: Errors list occurred during request processing
          items:
            type: string
      description: ''
...
  securitySchemes:
    API Key:
      name: API Key
      type: apiKey
      in: header
      description: Firebase token